Add separate kokkos_simd header
