+++ octotiger/common_kernel/kokkos_util.hpp
@@ -169,58 +169,4 @@ using normal_host_buffer = kokkos_host_array<T>;
 template <typename T>
 using normal_device_buffer = kokkos_device_array<T>;
 
-// =================================================================================================
-// SIMD types
-// =================================================================================================
-
-// defines HPX_COMPUTE_HOST_CODE and HPX_COMPUTE_DEVICE_CODE accordingly for the device passes
-// useful for picking the correct simd type!
-#include <hpx/config/compiler_specific.hpp> 
-// SIMD settings
-#if defined(OCTOTIGER_HAVE_STD_EXPERIMENTAL_SIMD)
-#include "octotiger/common_kernel/std_simd.hpp"
-#pragma message "Using std-experimental-simd SIMD types"
-#else
-#include <simd.hpp>
-using device_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::scalar>;
-using device_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::scalar>;
-#if !defined(HPX_COMPUTE_DEVICE_CODE) && !defined(OCTOTIGER_FORCE_SCALAR_KOKKOS_SIMD)
-#if defined(__VSX__)
-// NVCC does not play fair with Altivec! See another project with similar issues:
-// See https://github.com/dealii/dealii/issues/7328
-#ifdef __CUDACC__ // hence: Use scalar when using nvcc
-using host_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::scalar>;
-using host_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::scalar>;
-#warning "Using scalar SIMD types"
-#else // no nvcc: We can try to use the altivec vectorization
-#include <vsx.hpp>
-// TODO Actually test with a non-cuda kokkos build and/or clang
-// as it should get around the vsx problem
-using host_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::vsx>;
-using host_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::vsx>;
-#warning "Using VSX SIMD types"
-#endif
-#elif defined(__AVX512F__)
-#include <avx512.hpp>
-using host_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::avx512>;
-using host_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::avx512>;
-#warning "Using AVX512 SIMD types"
-#elif defined(__AVX2__) || defined(__AVX__)
-#include <avx.hpp>
-using host_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::avx>;
-using host_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::avx>;
-#warning "Using AVX SIMD types"
-#else
-#error "Could not detect any supported SIMD instruction set. Define OCTOTIGER_FORCE_SCALAR_KOKKOS_SIMD to continue anyway (or fix your arch flags if your platform supports AVX)!"
-#endif
-#else
-// drop in for nvcc device pass - is used on host side if FORCE_SCALAR_KOKKOS_SIMD is on
-// otherwise only used for compilation
-using host_simd_t = SIMD_NAMESPACE::simd<double, SIMD_NAMESPACE::simd_abi::scalar>;
-using host_simd_mask_t = SIMD_NAMESPACE::simd_mask<double, SIMD_NAMESPACE::simd_abi::scalar>;
-#warning "Using scalar SIMD types"
-#endif
-
-#endif
-
 #endif
